requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-telegram-bot==13.15
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
tornado==6.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
